Various warnings fixed (including 1 actual bug)

+++ sfml/test-typesetter.cpp
@@ -67,13 +67,13 @@ int main(int argc, char *argv[]) {
 
     buffy = &buffer;
 
-    for(int i=0;i<text.size();i++) {
+    for(int i=0;i<(int)text.size();i++) {
         wrapper.feed( FormattedCharacter( myFont, sf::Color(255,255,255), (uint32_t)text[i] ) );
     }
-    for(int i=0;i<moreText.size();i++) {
+    for(int i=0;i<(int)moreText.size();i++) {
         wrapper.feed( FormattedCharacter( mySecondFont, red, (uint32_t)moreText[i] ) );
     }
-    for(int i=0;i<evenMoreText.size();i++) {
+    for(int i=0;i<(int)evenMoreText.size();i++) {
         wrapper.feed( FormattedCharacter( myFont, white, (uint32_t)evenMoreText[i] ) );
     }
     wrapper.end();
